refactor: move page-related components to pages folder
